+++ app/Controllers/Invoice.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Controllers\Admin;
-use App\Controllers\BaseController;
+namespace App\Controllers;
 
 use App\Models\ModelInvoice;
 
-class Invoice extends BaseController{
+class Invoice extends BaseController
+{
     public function __construct()
     {
         $this->invoice = new ModelInvoice();
@@ -16,8 +16,8 @@ class Invoice extends BaseController{
         $data['invoice'] = $this->invoice->tampil_data();
         $data['total_items'] = $this->cart->totalItems();
         echo view('templates_admin/header');
-        echo view('templates_admin/sidebar',$data);
-        echo view('admin/invoice',$data);
+        echo view('templates_admin/sidebar', $data);
+        echo view('admin/invoice', $data);
         echo view('templates_admin/footer');
     }
 }
