--- app/Controllers/admin/DashboardAdmin.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Controllers\Admin;
-use App\Controllers\BaseController;
-
-class DashboardAdmin extends BaseController{
-    public function __construct()
-    {
-        $this->session = \Config\Services::session();
-        $this->session->start();
-    }
-    public function index()
-    {
-        echo view('templates_admin/header');
-        echo view('templates_admin/sidebar');
-        echo view('admin/dashboard');
-        echo view('templates_admin/footer');
-    }
-}
